> Fixed Ex. 2 from Assignment 4 (conversion of voltage to distance)

+++ p_4/e_2/components/distance_sensor/distance_sensor_reading.c
@@ -4,7 +4,7 @@ void make_distance_sensor_reading(
     distance_sensor_reading_t *reading,
     int adc_reading,
     int voltage_mv,
-    uint8_t distance_mm
+    int distance_mm
 ) {
     static uint64_t reading_id_counter = 0;
